1. Changed package name to  com.quantuminventions.customkeyboard
