+++ app/build.gradle
@@ -6,7 +6,7 @@ apply plugin: 'kotlin-kapt'
 android {
     compileSdkVersion 32
     defaultConfig {
-        applicationId "com.donbrody.customkeyboard"
+        applicationId "com.quantuminventions.customkeyboard.app"
         minSdkVersion 23
         targetSdkVersion 32
         versionCode 1
