+++ app/src/androidTest/java/com/quantuminventions/customkeyboard/app/ExampleInstrumentedTest.kt
@@ -1,0 +1,23 @@
+package com.quantuminventions.customkeyboard.app
+
+import androidx.test.ext.junit.runners.AndroidJUnit4
+
+import org.junit.Test
+import org.junit.runner.RunWith
+
+import org.junit.Assert.*
+
+/**
+ * Instrumented test, which will execute on an Android device.
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+@RunWith(AndroidJUnit4::class)
+class ExampleInstrumentedTest {
+    @Test
+    fun useAppContext() {
+        // Context of the app under test.
+        // val appContext = InstrumentationRegistry.getTargetContext()
+        // assertEquals("com.quantuminventions.customkeyboard.app", appContext.packageName)
+    }
+}
