+++ app/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.donbrody.customkeyboard">
+    package="com.quantuminventions.customkeyboard.app">
 
     <application
         android:allowBackup="true"
