+++ app/src/main/res/layout/activity_advanced_features.xml
@@ -82,7 +82,7 @@
         </LinearLayout>
     </ScrollView>
 
-    <com.kiranshaw.customdecimalkeyboard.components.keyboard.CustomisedKeyboardView
+    <com.quantuminventions.customkeyboard.components.keyboard.CustomisedKeyboardView
         android:id="@+id/customKeyboardView"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
