--- customdecimalkeyboard/src/main/java/com/kiranshaw/customdecimalkeyboard/components/expandableView/ExpandableState.kt
@@ -1,11 +1,0 @@
-package com.kiranshaw.customdecimalkeyboard.components.expandableView
-
-/**
- * Created by Don.Brody on 7/18/18.
- */
-enum class ExpandableState {
-    COLLAPSED,
-    COLLAPSING,
-    EXPANDED,
-    EXPANDING
-}
