--- customdecimalkeyboard/src/main/java/com/kiranshaw/customdecimalkeyboard/components/expandableView/ExpandableStateListener.kt
@@ -1,8 +1,0 @@
-package com.kiranshaw.customdecimalkeyboard.components.expandableView
-
-/**
- * Created by Don.Brody on 7/21/18.
- */
-interface ExpandableStateListener {
-    fun onStateChange(state: ExpandableState)
-}
