--- customdecimalkeyboard/src/main/java/com/kiranshaw/customdecimalkeyboard/components/keyboard/controllers/NumberDecimalKeyboardController.kt
@@ -1,21 +1,0 @@
-package com.kiranshaw.customdecimalkeyboard.components.keyboard.controllers
-
-import android.view.inputmethod.InputConnection
-
-/**
- * Created by Don.Brody on 7/20/18.
- */
-class NumberDecimalKeyboardController(inputConnection: InputConnection):
-    DefaultKeyboardController(inputConnection) {
-
-    override fun handleKeyStroke(c: Char) {
-        if (c == '.') {
-            // decimal numbers can only have one decimal point
-            if (!inputText().contains('.')) {
-                addCharacter(c)
-            }
-        } else {
-            addCharacter(c)
-        }
-    }
-}
