+++ customdecimalkeyboard/src/main/java/com/quantuminventions/customdecimalkeyboard/components/expandableView/ExpandableState.kt
@@ -1,0 +1,11 @@
+package com.quantuminventions.customkeyboard.components.expandableView
+
+/**
+ * Created by Don.Brody on 7/18/18.
+ */
+enum class ExpandableState {
+    COLLAPSED,
+    COLLAPSING,
+    EXPANDED,
+    EXPANDING
+}
