+++ customdecimalkeyboard/src/main/java/com/quantuminventions/customdecimalkeyboard/components/expandableView/ExpandableStateListener.kt
@@ -1,0 +1,8 @@
+package com.quantuminventions.customkeyboard.components.expandableView
+
+/**
+ * Created by Don.Brody on 7/21/18.
+ */
+interface ExpandableStateListener {
+    fun onStateChange(state: ExpandableState)
+}
