+++ customdecimalkeyboard/src/main/java/com/quantuminventions/customdecimalkeyboard/components/keyboard/controllers/NumberDecimalKeyboardController.kt
@@ -1,0 +1,21 @@
+package com.quantuminventions.customkeyboard.components.keyboard.controllers
+
+import android.view.inputmethod.InputConnection
+
+/**
+ * Created by Don.Brody on 7/20/18.
+ */
+class NumberDecimalKeyboardController(inputConnection: InputConnection):
+    DefaultKeyboardController(inputConnection) {
+
+    override fun handleKeyStroke(c: Char) {
+        if (c == '.') {
+            // decimal numbers can only have one decimal point
+            if (!inputText().contains('.')) {
+                addCharacter(c)
+            }
+        } else {
+            addCharacter(c)
+        }
+    }
+}
